Login Page works with a Database
Now in index.php. It'll always create the table for users whenever the
page loads. This is to ensure that the user table is already made. We
can decide if we wanted populated beforehand as well.  I'll take out the
error message before the demo as well.

Once the user logs in, a query to the DB will be made to make sure at
the user credentials are correct or not. If it is, the user will be
redirected to the dashboard.

UI and aesthetics will come after I get the register page up and
working. I know we discussed using templates from bootstrap for these
pages, but i don't know, maybe it'll be good practice to create it from
scratch?