User Registration Is done
A unique Userid will be generated whenever a new user is created. We
also check to see if there's an already existing user with the same user
name. If so, an error will be thrown and the user has to try again.

A successful register will take the user to the dashboard. I like the
idea of taking the user back to the login page so they can relog in, but
we can discuss that later.

Updated the CSS sheet, so that the fonts can be updated. Will be working
on some of the UI stuff to make it look better.